Move Stitch authentication logic into the libs folder

+++ api/apps/payments/views/linkpay.py
@@ -1,10 +1,7 @@
-import urllib.parse
 import uuid
 
-import requests
 import structlog
 from django.conf import settings
-from django.core.cache import cache
 from django.core.exceptions import ObjectDoesNotExist
 from rest_framework.response import Response
 from rest_framework.status import HTTP_400_BAD_REQUEST, HTTP_500_INTERNAL_SERVER_ERROR
@@ -14,8 +11,8 @@ from api.apps.payments.models import BankAccount, BankAccountToken
 from api.apps.payments.serializers.linkpay import PaymentAuthorizationSerializer, FetchUserTokenSerializer, \
     UnlinkAccountSerializer
 from api.apps.users.models import User
+from api.utils.libs.stitch.authentication import Authentication
 from api.utils.libs.stitch.errors import LinkPayError
-from api.utils.libs.stitch.helpers import generate_code_verifier_challenge_pair
 from api.utils.libs.stitch.linkpay.linkpay import LinkPay
 from api.utils.permissions import IsActiveUser
 
@@ -107,25 +104,13 @@ class CreatePaymentAuthorizationView(APIView):
                     content_type='application/json'
                 )
 
-            client_response = payment_authorization['clientPaymentAuthorizationRequestCreate']['authorizationRequestUrl']
-            scopes = urllib.parse.quote('openid transactions accounts balances accountholders offline_access paymentinitiationrequest')
-            client_id = settings.STITCH_CLIENT_ID
-            redirect_uri = settings.LINKPAY_REDIRECT_URI
-            code_verifier, code_challenge = generate_code_verifier_challenge_pair()
-            nonce, state = uuid.uuid4(), uuid.uuid4()
+            authorization_url = Authentication().generate_authorization_url(
+                base_url=payment_authorization['clientPaymentAuthorizationRequestCreate']['authorizationRequestUrl'],
+                scopes='openid transactions accounts balances accountholders offline_access paymentinitiationrequest'
+            )
 
-            session_data = {
-                'code_verifier': code_verifier
-            }
-
-            cache.set(state, session_data, 1800)
-
-            response = {
-                'url': f'{client_response}?client_id={client_id}&scope={scopes}&response_type=code&redirect_uri={redirect_uri}'
-                       f'&nonce={nonce}&state={state}&code_challenge={code_challenge}&code_challenge_method=S256'
-            }
             return Response(
-                data=response,
+                data=authorization_url,
                 content_type='application/json'
             )
 
@@ -141,46 +126,20 @@ class VerifyAndLinkUserAccount(APIView):
             authorization_code = serialized_data.validated_data['code']
             state = serialized_data.validated_data['state']
 
-            url = 'https://secure.stitch.money/connect/token'
-            previous_state = cache.get(state)
-            code_verifier = previous_state['code_verifier']
+            user_token_response = Authentication().get_user_token(
+                state=state, authorization_code=authorization_code
+            )
 
-            raw_data = {
-                'grant_type': 'authorization_code',
-                'client_id': f'{settings.STITCH_CLIENT_ID}',
-                'client_secret': f'{settings.STITCH_CLIENT_SECRET}',
-                'code': authorization_code,
-                'redirect_uri': f'{settings.LINKPAY_REDIRECT_URI}',
-                'code_verifier': code_verifier,
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer'
-            }
-            payload = urllib.parse.urlencode(raw_data)
-            headers = {
-                'Content-Type': 'application/x-www-form-urlencoded'
-            }
-
-            logger.debug('attempting to fetch user token using authorization code')
-
-            try:
-                response = requests.request(
-                    'POST', url, headers=headers, data=payload
-                )
-                response.raise_for_status()
-            except requests.exceptions.RequestException as e:
-                error = f'could not obtain user token: {str(e)}'
-                logger.error(error)
-
+            if 'error' in user_token_response:
                 return Response(
-                    data={'error': error},
+                    data=user_token_response,
                     status=HTTP_500_INTERNAL_SERVER_ERROR,
                     content_type='application/json'
                 )
 
-            user_token = response.json()
-
             try:
                 linked_user = request.user
-                account_details = fetch_linked_account_details(user_token['access_token'])
+                account_details = fetch_linked_account_details(user_token_response['access_token'])
 
                 if (linked_user.get_full_name() == account_details['accountHolder']['fullName']) and \
                         linked_user.identification_number == account_details['accountHolder']['identifyingDocument']['number']:
@@ -193,7 +152,7 @@ class VerifyAndLinkUserAccount(APIView):
                             content_type='application/json'
                         )
 
-                    save_linked_account_details(request.user, account_details, user_token)
+                    save_linked_account_details(request.user, account_details, user_token_response)
 
                     return Response(
                         data={'success': 'Account successfully linked'},
@@ -246,32 +205,22 @@ class UnlinkUserAccount(APIView):
                     content_type='application/json'
                 )
 
-            request_body = {
-                'client_id': settings.STITCH_CLIENT_ID,
-                'client_secret': settings.STITCH_CLIENT_SECRET,
-                'token': refresh_token,
-                'token_type_hint': 'refresh_token',
-            }
+            token_revoked = Authentication().revoke_token(
+                token=refresh_token, token_type='refresh_token'
+            )
 
-            url = 'https://secure.stitch.money/connect/revocation'
-            payload = urllib.parse.urlencode(request_body)
-            headers = {
-                'Content-Type': 'application/x-www-form-urlencoded'
-            }
-
-            try:
-                response = requests.request(
-                    'POST', url, headers=headers, data=payload
-                )
-                response.raise_for_status()
+            if token_revoked:
                 logger.info('Refresh token revoked successfully on Stitch')
-            except requests.exceptions.RequestException as e:
-                logger.error(f'could not revoke refresh token on Stitch: {str(e)}')
+                linked_account.delete()
+                logger.info('Account records and token successfully deleted')
 
-            linked_account.delete()
-            logger.info('Account records and token successfully deleted')
+                return Response(
+                    data={'success': 'Account successfully unlinked'},
+                    content_type='application/json'
+                )
 
             return Response(
-                data={'success': 'Account successfully unlinked'},
+                data={'error': 'Account not unlinked'},
+                status=HTTP_400_BAD_REQUEST,
                 content_type='application/json'
             )
